refactor: TodoError의 소스파일 분리 및 불필요한 import 삭제
