+++ ProjectManager/ProjectManager/Extension/NameSpace.swift
@@ -5,8 +5,6 @@
 //  Created by 애쉬 on 2023/01/19.
 //
 
-import Foundation
-
 enum HeaderViewValue {
     static let identifier = "ListHeaderView"
 }
@@ -62,22 +60,4 @@ enum AlertMenu {
     static let toTodo = "Move To TODO"
     static let toDoing = "Move To DOING"
     static let toDone = "Move To DONE"
-}
-
-enum TodoError {
-    case loadError
-    
-    var title: String {
-        switch self {
-        case .loadError:
-            return "데이터 로드 오류"
-        }
-    }
-    
-    var message: String {
-        switch self {
-        case .loadError:
-            return "해당 항목의 데이터를 가져올 수 없습니다."
-        }
-    }
 }
